code reorg, data, handlers and subscriber modules added to the crate
